# backend/app/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Mis Eventos API
  version: 1.0.0
  description: |
    API para gestionar cuentas, eventos, sesiones y registros de asistentes.
    Las rutas utilizan el prefijo /api/v1.
servers:
  - url: /api/v1
    description: Servidor de desarrollo local
...
components:
  securitySchemes:
    AccessTokenCookie:
      type: apiKey
      in: cookie
      name: access_token
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: validation_error
            message:
              type: string
              example: The request contains invalid fields.
...
    User:
      type: object
      required: [id, email, first_name, last_name]
      properties:
        id:
          type: integer
          example: 7
        email:
          type: string
          format: email
          example: user@example.com
        first_name:
          type: string
          example: Alex
        last_name:
          type: string
          example: Rivera
...
paths:
...
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, first_name, last_name]
              properties:
                email:
                  type: string
                  format: email
                  example: attendee@example.com
                password:
                  type: string
                  format: password
                  example: correct-horse-battery
                first_name:
                  type: string
                  example: Alex
                last_name:
                  type: string
                  example: Rivera
      responses:
        "201":
          description: Account created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
        "400":
          description: Invalid request or validation error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Account already exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/login:
    post:
      tags: [Authentication]
      summary: Authenticate and set the access-token cookie
      description: |
        The response sets an HttpOnly access_token cookie.
        Its Secure flag depends on the application configuration.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: attendee@example.com
                password:
                  type: string
                  format: password
                  example: correct-horse-battery
      responses:
        "200":
          description: Authentication succeeded; access-token cookie is set.
          headers:
            Set-Cookie:
              description: HttpOnly access_token cookie.
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
        "400":
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid credentials.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Authentication unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/me:
    get:
      tags: [Authentication]
      summary: Get the current user
      security:
        - AccessTokenCookie: []
      responses:
        "200":
          description: Current user returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
        "401":
          description: Authentication required or token invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Authentication unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...